fastapi_rvc/pyproject.toml
[tool.poetry]
name = "rvc"
version = "0.3.5"
description = "An easy-to-use Voice Conversion framework based on VITS."
authors = ["Ftps <ftpsflandre@gmail.com>"]
readme = "README.md"

[tool.poetry.urls]
github = "https://github.com/RVC-Project/Retrieval-based-Voice-Conversion"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.0"
# fairseq = "^0.12.2"
fairseq = { git = "https://github.com/Tps-F/fairseq.git", branch = "main" }
soundfile = "^0.12.1"
librosa = "^0.10.1"
praat-parselmouth = "^0.4.3"
pyworld = "^0.3.4"
torchcrepe = "^0.0.22"
av = "^11.0.0"
faiss-cpu = "^1.7.4"
python-dotenv = "^1.0.0"
pydub = "^0.25.1"
click = "^8.1.7"
tensorboardx = "^2.6.2.2"
poethepoet = "^0.24.4"
uvicorn = { version = "^0.26.0", optional = true }
fastapi = "^0.109.0"
python-multipart = { version = "^0.0.6", optional = true }
numba = "0.59.0rc1"

[tool.poetry.extras]
api = ["uvicorn", "fastapi"]

[tool.poetry.scripts]
rvc = "rvc.wrapper.cli.cli:main"

[tool.poe.tasks]
rvc-api = "uvicorn rvc.wrapper.api.api:app --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
